Move AgentConf to agent.py
