+++ integration/test/test_template.py.in
@@ -111,7 +111,7 @@ class TestIntegration_@test_name@(unittest.TestCase):
             freq_sticker = geopm_test_launcher.geopmread("CPUINFO::FREQ_STICKER board 0")
             agent_conf_dict = {'FREQ_MIN': freq_min,
                                'FREQ_MAX': freq_sticker}
-            agent_conf = geopmpy.io.AgentConf(cls._agent_conf_path,
+            agent_conf = geopmpy.agent.AgentConf(cls._agent_conf_path,
                                               'energy_efficient',
                                               agent_conf_dict)
             # Create the test launcher with the above configuration
